Adding Income (#117)

+++ backend/migrations/0009_add_budget_income_props.sql
@@ -1,5 +1,5 @@
 ALTER TABLE budgets 
-ADD paycheck_amount int8 NULL;
+ADD paycheck_amount FLOAT NULL;
 
 ALTER TABLE budgets 
-ADD additional_income_amount int8 NULL;
+ADD additional_income_amount FLOAT NULL;
